Adds support for NO_DELETE_INDEX (patch)

+++ index.js
@@ -1,5 +1,6 @@
 const axios = require('axios')
 const path = require('path')
+const sleep = require('then-sleep')
 const addIndex = require('./lib/add-index')
 const deleteIndex = require('./lib/delete-index')
 const repackPolitician = require('./lib/repack-politician')
@@ -11,14 +12,21 @@ if (env) {
   const envFilePath = path.resolve(process.cwd(), env)
   logger('info', ['index', 'loading environment', env])
   require('dotenv').config({ path: envFilePath })
+} else if (process.env.NODE_ENV !== 'production') {
+  logger('info', ['index', 'loading environment', '.env'])
+  require('dotenv').config()
 } else {
   logger('warn', ['index', 'no environment loaded'])
 }
 
 async function indexPoliticians () {
   const { data } = await axios.get(process.env.SOURCE_URL)
-  const msg = await deleteIndex()
-  logger('info', ['index', 'indexPoliticians', 'index deleted', JSON.stringify(msg)])
+  if (!process.env.NO_DELETE_INDEX) {
+    const msg = await deleteIndex()
+    logger('info', ['index', 'indexAnsatte', 'index deleted', JSON.stringify(msg)])
+  } else {
+    logger('info', ['index', 'indexAnsatte', 'no delete index'])
+  }
   let list = data.map(repackPolitician).map(prepareIndex)
   logger('info', ['index', 'indexPoliticians', 'politicians to index', list.length])
   let success = 0
@@ -33,6 +41,7 @@ async function indexPoliticians () {
       } catch (error) {
         fail++
       }
+      await sleep(100)
       await next()
     } else {
       logger('info', ['index', 'indexPoliticians', 'finished', 'success', success, 'fail', fail])
